refactor: extract LogDirective to its own file
