+++ src/core/assistant-message/index.ts
@@ -1,1 +1,2 @@
 export { type AssistantMessageContent, parseAssistantMessage } from "./parse-assistant-message"
+export { type LogDirective } from "./directives/log-directive"
